requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
diff-match-patch==20241021; python_version >= '3.7'
django==4.2.16; python_version >= '3.8'
django-import-export==4.3.1; python_version >= '3.9'
djangorestframework==3.15.2; python_version >= '3.8'
psycopg2==2.9.10; python_version >= '3.8'
sqlparse==0.5.2; python_version >= '3.8'
tablib==3.7.0; python_version >= '3.9'
